Import latest DLABridging:work and DLABCaptureManager:work into DLAB SPM (#3)
* Initial plan

* Import DLABridging:work and DLABCaptureManager:work changes

- Update DeckLink API SDK to version 15.3
  - Add DeckLinkAPI_v15_2.h and DeckLinkAPIDispatch_v15_2.cpp (new)
  - Update DeckLinkAPI.h, DeckLinkAPIConfiguration.h,
    DeckLinkAPIDispatch.cpp, DeckLinkAPITypes.h, DeckLinkAPIVersion.h
- Update DLABConstants.h per SDK 15.3
- Add DataSpace support to DLABAncillaryPacket
- Add unified ancillary packet handlers to DLABDevice
  (DLABDevice.h, DLABDevice+Internal.h, DLABDevice+Input.mm,
   DLABDevice+Output.mm, DLABDevice.mm)
- Fix enum bitwise warnings + ISO C++ warnings in DLABVideoConverter.mm
- Fix hash XOR operand type in DLABVideoSetting.mm
- Fix QueryInterface cast in DLABDeckControlStatusCallback.mm
- CaptureManager.swift: Swift 6 sendable, prewarm, ancillary handler
- CaptureWriter.swift: Sendable, prewarm, locks
- CaptureVideoPreview.swift: preview safety, relayout, cache reset
- VideoSampleBufferHelper.swift: reset sample cache
- Update copyright year 2025→2026 for all files
- Update README.md copyright year 2025→2026



* Address PR review feedback


* Address additional PR review comments


* Address ancillary packet review feedback


* Use explicit default ancillary dataSpace


* Address capture review feedback


* Fix internal frameDuration doc wording


* Address latest review comments


* Fix local DLABridging header imports


* Fix vImage matrix type initializer


---------

+++ Sources/DLABridging/src/DLABVideoSetting.mm
@@ -3,7 +3,7 @@
 //  DLABCore
 //
 //  Created by Takashi Mochizuki on 2017/08/26.
-//  Copyright © 2017-2025 MyCometG3. All rights reserved.
+//  Copyright © 2017-2026 MyCometG3. All rights reserved.
 //
 
 /* This software is released under the MIT License, see LICENSE.txt. */
@@ -343,7 +343,7 @@ NS_INLINE BOOL checkPixelFormat(BMDPixelFormat dlPixelFormat, OSType cvPixelForm
 // public hash - NSObject
 - (NSUInteger) hash
 {
-    NSUInteger value = (NSUInteger)(_widthW^_heightW) ^ (NSUInteger)(_displayMode^_pixelFormatW);
+    NSUInteger value = ((NSUInteger)_widthW ^ (NSUInteger)_heightW) ^ ((NSUInteger)_displayMode ^ (NSUInteger)_pixelFormatW);
     return value;
 }
